Anonymous Users can now save lists
Issue was that we weren’t adding the new user to the list of users of
the new list. That meant that when web.views.list_list was called that
request.user.list_set.all() returned an empty list and nothing was
rendered to the user.

Tested it real quick and it seems to work now.